Fixed the inconsistent navbars on the admin pages
