+++ .editorconfig
@@ -4,7 +4,7 @@ root = true
 [*]
 charset = utf-8
 indent_style = tab
-indent_size = 3
+tab_width = 3
 insert_final_newline = true
 trim_trailing_whitespace = true
