+++ packages/frontend/src/pages/admin/map-builder.tsx
@@ -1,22 +1,23 @@
 import React, { useState } from 'react';
 import { Card } from 'react-bootstrap';
 import { FaMap } from 'react-icons/fa';
-import { useNavigate } from 'react-router';
 
 import { RoomBuilder } from '../../components/room-builder';
+
+import { AdminNavbar } from './admin-navbar';
+
 
 const MENU_BAR_HEIGHT = 56;
 
 export const AdminMapBuilder: React.FC = () => {
 	const [selectedFloor, setSelectedFloor] = useState(0);
-	const navigate = useNavigate();
 
 	const handleFloorChange = (floor: number) => {
 		setSelectedFloor(floor);
 	};
 
 	return (
-		<div style={{ position: 'relative', minHeight: '100vh', background: '#181a28' }}>
+		<div className="admin-layout">
 			{/* Fixed Menu Bar */}
 			<div
 				style={{
@@ -25,26 +26,10 @@ export const AdminMapBuilder: React.FC = () => {
 					left: 0,
 					width: '100%',
 					height: MENU_BAR_HEIGHT,
-					background: '#222',
-					color: '#fff',
-					display: 'flex',
-					alignItems: 'center',
 					zIndex: 2000,
-					boxShadow: '0 2px 8px #000a',
-					padding: '0 2rem',
 				}}
 			>
-				<span style={{ fontWeight: 700, marginRight: 32, fontSize: 18, cursor: 'pointer' }} onClick={() => navigate('/admin')}>
-					Stargate Evolution Admin
-				</span>
-				<a style={{ color: '#fff', marginRight: 24, textDecoration: 'none', cursor: 'pointer' }} onClick={() => navigate('/admin')}>Overview</a>
-				<a style={{ color: '#fff', marginRight: 24, textDecoration: 'none', cursor: 'pointer', fontWeight: 600 }} onClick={() => navigate('/admin/map')}>Map Builder</a>
-				<a style={{ color: '#fff', marginRight: 24, textDecoration: 'none', cursor: 'pointer' }} onClick={() => navigate('/admin/users')}>Users</a>
-				<a style={{ color: '#fff', marginRight: 24, textDecoration: 'none', cursor: 'pointer' }} onClick={() => navigate('/admin/characters')}>Characters</a>
-				<a style={{ color: '#fff', marginRight: 24, textDecoration: 'none', cursor: 'pointer' }} onClick={() => navigate('/admin/rooms')}>Room Templates</a>
-				<a style={{ color: '#fff', marginRight: 24, textDecoration: 'none', cursor: 'pointer' }} onClick={() => navigate('/admin/technologies')}>Technologies</a>
-				<div style={{ flex: 1 }} />
-				<a style={{ color: '#fff', textDecoration: 'none', cursor: 'pointer', fontWeight: 500 }} onClick={() => navigate('/')}>Back to Game</a>
+				<AdminNavbar />
 			</div>
 
 			{/* Content below menu bar, add top margin to avoid overlap */}
